+++ src/Sulu/Bundle/SecurityBundle/Entity/User.php
@@ -190,7 +190,7 @@ class User extends BaseUser
     {
         $userSettingValues = [];
         foreach ($this->userSettings as $userSetting) {
-            $userSettingValues[$userSetting->getKey()] = json_decode($userSetting->getValue());
+            $userSettingValues[$userSetting->getKey()] = json_decode($userSetting->getValue(), true);
         }
 
         return $userSettingValues;